fix(pop-boot): improve pop-boot postinst
This ensures that pop-boot's postinst doesn't run if boot configuration is
already present on the system. Otherwise, it may overwrite the system boot
options with new ones, which isn't good. It also makes these fields much
easier to modify and maintain, and prints a helpful message if it skipped
the postinst because the configuration was already found (along with that
config location).